requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
pydantic>=2.4.2
google-cloud-speech>=2.27.0
google-cloud-storage>=2.18.0
google-cloud-texttospeech>=2.18.0
google-cloud-language>=2.14.0
Flask>=3.0.2
gunicorn>=20.1.0

